Fix build errors
* Move version into constant in main.go
* Move usage() func to main.go
