--- version.go
@@ -1,5 +1,0 @@
-package main
-
-func versionString() string {
-	return "gethub 0.1.3"
-}
